BelowZero - Extracted the common behaviors of RecipeData and TechData into an abstract base class
